+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\components"
+  ],
+  "SelectedNode": "\\components\\index.js",
+  "PreviewInSolutionExplorer": false
+}
